@@ -10,25 +10,31 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.0.8",
-    "@angular/cdk": "^17.0.4",
-    "@angular/common": "^17.0.0",
-    "@angular/compiler": "^17.0.0",
-    "@angular/core": "^17.0.0",
-    "@angular/fire": "^17.0.1",
-    "@angular/forms": "^17.0.0",
-    "@angular/material": "^17.0.4",
-    "@angular/material-moment-adapter": "^17.1.2",
-    "@angular/platform-browser": "^17.0.0",
-    "@angular/platform-browser-dynamic": "^17.0.0",
-    "@angular/router": "^17.0.0",
+    "@angular/animations": "^18.2.1",
+    "@angular/cdk": "^18.2.1",
+    "@angular/common": "^18.2.1",
+    "@angular/compiler": "^18.2.1",
+    "@angular/core": "^18.2.1",
+    "@angular/fire": "^18.0.1",
+    "@angular/forms": "^18.2.1",
+    "@angular/material": "^18.2.1",
+    "@angular/material-moment-adapter": "^18.2.1",
+    "@angular/platform-browser": "^18.2.1",
+    "@angular/platform-browser-dynamic": "^18.2.1",
+    "@angular/router": "^18.2.1",
     "@apollo/client": "^3.9.11",
+    "@fortawesome/angular-fontawesome": "^0.15.0",
+    "@fortawesome/fontawesome-svg-core": "^6.6.0",
+    "@fortawesome/free-brands-svg-icons": "^6.6.0",
+    "@fortawesome/free-solid-svg-icons": "^6.6.0",
     "angular-calendar": "^0.31.0",
     "angular-star-rating": "^7.0.0",
-    "apollo-angular": "^6.0.0",
+    "apollo-angular": "^7.1.2",
     "bootstrap": "^5.3.2",
+    "bootstrap-icons": "^1.11.3",
     "css-star-rating": "^1.3.1",
     "date-fns": "^3.3.1",
+    "emailjs-com": "^3.2.0",
     "firebase": "^10.12.3",
     "firebase-tools": "^13.13.1",
     "graphql": "^16",
@@ -37,16 +43,17 @@
     "ng-otp-input": "^1.9.3",
     "ngx-countdown": "^17.1.1",
     "ngx-loading": "^17.0.0",
-    "ngx-mat-timepicker": "^17.1.0",
+    "ngx-mat-timepicker": "^18.0.0",
     "ngx-scrollbar": "^13.0.3",
+    "ngx-sharebuttons": "^15.0.2",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
-    "zone.js": "~0.14.2"
+    "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.0.7",
-    "@angular/cli": "^17.0.7",
-    "@angular/compiler-cli": "^17.0.0",
+    "@angular-devkit/build-angular": "^18.2.1",
+    "@angular/cli": "^18.2.1",
+    "@angular/compiler-cli": "^18.2.1",
     "@types/crypto-js": "^4.2.2",
     "@types/jasmine": "~5.1.0",
     "jasmine-core": "~5.1.0",
@@ -55,6 +62,6 @@
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "typescript": "~5.2.2"
+    "typescript": "~5.4.0"
   }
 }
