@@ -13,7 +13,7 @@ import { MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { CreateNewPasswordComponent } from './auth/create-new-password/create-new-password.component';
 import { ForgotPasswordComponent } from './auth/forgot-password/forgot-password.component';
 import { ResetPasswordComponent } from './auth/reset-password/reset-password.component';
-import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
+import { HTTP_INTERCEPTORS, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { ErrorInterceptor } from './core/interceptors/error.interceptor';
 import { AuthGuard } from './core/guard/auth.guard';
 import { AuthService } from './shared/services/auth.service';
@@ -28,40 +28,37 @@ import { AngularFireMessagingModule } from '@angular/fire/compat/messaging';
 import { environment } from '../enviroments/enviroment';
 
 @NgModule({
-  declarations: [
-    AppComponent,
-    LoginComponent,
-    SignupComponent,
-    ForgotPasswordComponent,
-    CreateNewPasswordComponent,
-    ResetPasswordComponent,
-    ContactUsComponent,
-  ],
-
-  imports: [
-    FormsModule,
-    ReactiveFormsModule,
-    BrowserModule,
-    AppRoutingModule,
-    SharedModule,
-    MaterialModule,
-    BrowserAnimationsModule,
-    NgOtpInputModule,
-    HttpClientModule,
-    GraphQLModule,
-    NgxLoadingModule.forRoot({}),
-    AngularFireMessagingModule,
-    AngularFireModule.initializeApp(environment.firebase)
-  ],
-  providers: [
-    AuthGuard,
-    AuthService,
-    ApolloService,
-    ToastMessageService,
-    { provide: MAT_FORM_FIELD_DEFAULT_OPTIONS, useValue: { float: 'never' } },
-    { provide: MAT_DIALOG_DATA, useValue: {} },
-    { provide: HTTP_INTERCEPTORS, useClass: ErrorInterceptor, multi: true },
-  ],
-  bootstrap: [AppComponent],
+    declarations: [
+        AppComponent,
+        LoginComponent,
+        SignupComponent,
+        ForgotPasswordComponent,
+        CreateNewPasswordComponent,
+        ResetPasswordComponent,
+        ContactUsComponent,
+    ],
+    bootstrap: [AppComponent],
+    imports: [FormsModule,
+        ReactiveFormsModule,
+        BrowserModule,
+        AppRoutingModule,
+        SharedModule,
+        MaterialModule,
+        BrowserAnimationsModule,
+        NgOtpInputModule,
+        GraphQLModule,
+        NgxLoadingModule.forRoot({}),
+        AngularFireMessagingModule,
+        AngularFireModule.initializeApp(environment.firebase)],
+    providers: [
+        AuthGuard,
+        AuthService,
+        ApolloService,
+        ToastMessageService,
+        { provide: MAT_FORM_FIELD_DEFAULT_OPTIONS, useValue: { float: 'never' } },
+        { provide: MAT_DIALOG_DATA, useValue: {} },
+        { provide: HTTP_INTERCEPTORS, useClass: ErrorInterceptor, multi: true },
+        provideHttpClient(withInterceptorsFromDi()),
+    ]
 })
 export class AppModule { }
