@@ -18,13 +18,13 @@ export class CaseDiaryLoginComponent {
 
   constructor(private _formBuilder: FormBuilder, private _toastMessage: ToastMessageService, private _apolloService: ApolloService,
     private _router: Router) {
-    let userData = JSON.parse(localStorage.getItem('userData')!);
+    let userData = JSON.parse(sessionStorage.getItem('userData')!);
     this.caseDiaryForm = this._formBuilder.group({
       userName: new FormControl('', [Validators.required]),
       password: new FormControl('', [Validators.required, Validators.minLength(10)]),
       lawyerId: new FormControl(userData._id)
     });
-    let isCaseDiaryLogin = JSON.parse(localStorage.getItem('isCaseDiaryLogin')!);
+    let isCaseDiaryLogin = JSON.parse(sessionStorage.getItem('isCaseDiaryLogin')!);
     if (isCaseDiaryLogin) {
       this._router.navigate(['lawyer/case-diary/cases']);
     }
@@ -37,7 +37,7 @@ export class CaseDiaryLoginComponent {
           if (data.data != null) {
             if (data.data.signIn.status == 200) {
               this._router.navigate(['lawyer/case-diary/cases']);
-              localStorage.setItem('isCaseDiaryLogin', JSON.stringify(true));
+              sessionStorage.setItem('isCaseDiaryLogin', JSON.stringify(true));
             }
             else {
               this._toastMessage.error(data.data.signIn.message);
