@@ -42,7 +42,7 @@ export class CaseDiarySignupComponent {
 
   register() {
     if (this.caseDiarySignUpForm.valid) {
-      let userData = localStorage.getItem('userData');
+      let userData = sessionStorage.getItem('userData');
       let parsedData = userData ? JSON.parse(userData) : {};
       let data = {
         lawyerId: parsedData._id,
