@@ -2,31 +2,36 @@ import { Component, Input } from '@angular/core';
 import { Router } from '@angular/router';
 import { Location } from '@angular/common';
 import { AuthService } from '../../services/auth.service';
-import { P } from '@angular/cdk/keycodes';
 import { imageUrl } from '../../../graphql.module';
 import { HttpClient } from '@angular/common/http';
+import { Subject, Subscription, takeUntil, timeout } from 'rxjs';
+
 
 @Component({
   selector: 'app-header',
   templateUrl: './header.component.html',
-  styleUrl: './header.component.scss'
+  styleUrl: './header.component.scss',
 })
 export class HeaderComponent {
   userType: string = "";
   userImage: string = "";
   notifications: any = [];
+  sub$: Subscription;
 
   @Input() menuName: string = "";
   @Input() searchStyle = { width: '0px', display: 'none' };
   @Input() searchIcon = { width: 'auto', display: 'block' };
   @Input() colConfig: string = "col-lg-8 col-md-8";
+  onDestroy$: Subject<void> = new Subject();
 
   constructor(private _router: Router, private _location: Location, private _authService: AuthService, private _http: HttpClient) {
+    this.sub$ = this._authService.profileImageSubject.asObservable()
+      .pipe(takeUntil(this.onDestroy$))
+      .subscribe((data: any) => {
+        this.userImage = data;
+      });
     this.getNotificationList();
     this.userType = this._router.url.split('/')[1];
-    this._authService.profileImageSubject.subscribe(data => {
-      this.userImage = data; console.log(data, 'userImage')
-    });
   }
 
   redirectToProfile() {
@@ -68,17 +73,17 @@ export class HeaderComponent {
   }
 
   getNotificationList() {
-    let userData = JSON.parse(localStorage.getItem('userData')!);
+    let userData = JSON.parse(sessionStorage.getItem('userData')!);
     this._http.get(imageUrl() + `/notifications/${userData._id}/${userData.userType.toLowerCase()}`).subscribe((data: any) => {
       if (data != null) {
         if (data.status == 200) {
           this.notifications = data.data;
-        }
-        else {
-          this.notifications = {
-            title: "No Notifications",
-            notificationType: "noData",
-            body: "No Notifications found !!",
+          if (this.notifications.length == 0) {
+            this.notifications = [{
+              title: "No Notifications",
+              notificationType: "noData",
+              body: "No Notifications found !!",
+            }]
           }
         }
       }
@@ -94,5 +99,10 @@ export class HeaderComponent {
   openNoification() {
     let el = document.getElementById('openNotifications') as HTMLElement;
     el.click();
+  }
+
+  ngOnDestroy() {
+    this.onDestroy$.next();
+    this.onDestroy$.complete();
   }
 }
