@@ -15,6 +15,14 @@ import { SafeHTMLPipe } from './pipe/safe-html.pipe';
 import { PreventKeyboardEventsDirective } from './directives/prevent-keyboard-events.directive';
 import { FormsModule } from '@angular/forms';
 
+import { ShareButtons } from 'ngx-sharebuttons/buttons';
+import { shareIcons } from 'ngx-sharebuttons/icons';
+import { SubscriptionPlanComponent } from './component/subscription-plan/subscription-plan.component';
+import { QrModalComponent } from './component/qr-modal/qr-modal.component';
+import { HighlightOnSearchPipe } from './pipe/highlight-on-search.pipe';
+import { LandingPageComponent } from './component/landing-page/landing-page.component';
+import { RouterModule } from '@angular/router';
+
 
 @NgModule({
   declarations: [
@@ -26,14 +34,20 @@ import { FormsModule } from '@angular/forms';
     HighlighterPipe,
     PaginatorComponent,
     SafeHTMLPipe,
-    PreventKeyboardEventsDirective
+    PreventKeyboardEventsDirective,
+    SubscriptionPlanComponent,
+    QrModalComponent,
+    HighlightOnSearchPipe,
+    LandingPageComponent
   ],
   imports: [
     CommonModule,
     MaterialModule,
     NgScrollbarModule,
     MatChipsModule,
-    FormsModule
+    FormsModule,
+    ShareButtons,
+    RouterModule
   ],
   exports: [
     HeaderComponent,
@@ -42,8 +56,12 @@ import { FormsModule } from '@angular/forms';
     HighlighterPipe,
     PaginatorComponent,
     SafeHTMLPipe,
-    PreventKeyboardEventsDirective
+    PreventKeyboardEventsDirective,
+    ShareButtons,
+    SubscriptionPlanComponent,
+    QrModalComponent,
+    HighlightOnSearchPipe
   ],
-  providers: [AuthService]
+  providers: [AuthService, shareIcons()]
 })
 export class SharedModule { }
